[rfct] split decorators into topical modules
